Make all integration tests into unit tests
Put them in the module at the lowest level of the stratis structure that
they test.

In some cases, simplify the tests so that they can be placed at a lower
level, i.e., in pool instead of engine module. Originally, there was some
effort to only invoke trait methods for tests, causing the engine to
be created even if the test was really just on an individual pool.

Make all strat modules private.

Add a few dead code annotations. Ultimately, if these persist, the code
itself should be eliminated.

Remove a few unnecessary methods that had been introduced solely for testing.

Pass the correct size to ThinDev::setup() so that it doesn't always fail,
which would mask any bugs that it was trying to check for.

Remove test_teardown. Pools are torn down as a matter of course in all the
tests, so this test is pointless.

Update documentation.
