+++ src/engine/strat_engine/tests/mod.rs
@@ -2,11 +2,10 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
+extern crate env_logger;
+extern crate log;
+pub extern crate tempdir;
 
-pub mod blockdev_tests;
-pub mod dm_tests;
-pub mod filesystem_tests;
-pub mod pool_tests;
-pub mod setup_tests;
-pub mod simple_tests;
-pub mod logger;
+mod logger;
+pub mod loopbacked;
+pub mod real;
